fixed sudoku brach to match updates on main

+++ gidd/loss.py
@@ -56,8 +56,8 @@ class GiddLoss(Loss):
         c_t = t_gamma.sqrt() * t1m_gamma.sqrt() * B
         c_t_prime = (gamma / 2) * (1 - 2 * t) / (t * t1m) * c_t
 
-        is_mask = (z_t == self.mask_id).float()
-        is_x = (z_t == input_ids).float()
+        is_mask = (z_t == self.mask_id).to(t.dtype)
+        is_x = (z_t == input_ids).to(t.dtype)
 
         alpha_ratio = -1 / (1 - t) - c_t_prime / (1 + c_t)
         N = self.vocab_size_semantically - 1
